Fixed global module functionality for Windows.
cmd-shim appends ".cmd" to the destination path when called. Yarn calls cmd-shim with the destination parameter `<package-name>.cmd`, resulting in an improperly formed shim: `<package-name>.cmd.cmd` referencing the wrong source: `<package-name>.cmd`. If the ".cmd" extension is removed from dest prior to calling cmd-shim, the result is `<package-name>.cmd` and `<package-name>`, respectively. This reliably fixes the behavior of global add and global remove in Windows.